somewhat consolidate code on scraping benchmark stats
